BREAKING CHANGE - reworking of models
This includes a lot of churn, mainly around removing the onfido_id as
the primary key and making it unique instead. This came out of some
other work where it proved impossible to save onfido models using a
generic foreign key because it was using a char field not an int. A
lesson learnt.