Add support for refreshing objects from API
The BaseModel now contains a `pull` method which will refresh the
internal properties from the API, and update the raw field.

I have also pulled the create_FOO functions out of the api module into
a helpers module to make dependencies easier to manage. As this breaks
the interface I will bump the version and re-release.