Add fetch method to BaseModel
This commit splits the previous pull method into two parts - fetch,
which calls the API, but does **not** save the object, and pull, which
calls fetch and then saves the model. This is useful for situations
where you want to inspect the remote object, but without updating the
local version.