Fix Event model fields (unique constraint)
The Event model was inheriting a unique=True constraint on the
Onfido ID field which is invalid for events. I've shifted the fields
around and removed the constraint. Additional change is to rename
the created_at field back to its original (and correct as per the
API) name of completed_at.