Add support for native Postgres JSONField (drop support for Django 1.8)
This commit drops the custom JSONField, which backed on to a TextField,
replacing it with the new django.contrib.postgres.fields.jsonb.JSONField.

Doing this has a number of implications, first of which is that support
for Django 1.8 and below is officially dropped from this commit on. The
migration of the field is a simple ALTER statement, coverting the
textfield to a jsonb field in postgres. Fortunately it appears to do
this in situ - all of the content is converted over from text without
error (it's been tried on a live database with 1,000s rows of JSON). In
the event that this doesn't work, a more 'correct' migration script is
in the comments inside the 0010 migration, which includes creating a
new field, copying the data over, deleting the old field.

The other significant change is that this project is now bound to Postgres.
Which in turn means that the tests require postgres to run (unless you feel
like mocking everything out).