Add management command to pull (sync) objects
The `onfido_sync` management command can be used to call the pull()
method on the entire dataset of Applicant, Check or Report objects.
It takes a single positional arg - 'check' or 'report', and has two
options - filter and exclude - both of which take multiple space-
separated args which can be used to manage the queryset that is used.

Examples:

    $ ./manage.py onfido_sync check
    $ ./manage.py onfido_sync report
    $ ./manage.py onfido_sync check --filter complete
    $ ./manage.py onfido_sync check --exclude complete