Add `received_at` DateTimeField to Event
The Event model already has a .completed_at DateTimeField used
as a timestamp but the time stored in it is that of the stringified
time we receive in the JSON webhook payload from Onfido.

Previous to this commit, we were not recording the actual time the
event hit our server endpoint and so did not have a real source of truth
for the actual time we received the event. This commit remedies that.